Moved redirect_to_login to utils.py
