+++ django/contrib/admin/sites.py
@@ -231,7 +231,7 @@ class AdminSite:
                     return HttpResponseRedirect(index_path)
                 # Inner import to prevent django.contrib.admin (app) from
                 # importing django.contrib.auth.models.User (unrelated model).
-                from django.contrib.auth.views import redirect_to_login
+                from django.contrib.auth.utils import redirect_to_login
 
                 return redirect_to_login(
                     request.get_full_path(),
