+++ django/contrib/auth/utils.py
@@ -1,0 +1,21 @@
+from urllib.parse import urlparse, urlunparse
+
+from django.conf import settings
+from django.contrib.auth import REDIRECT_FIELD_NAME
+from django.http import HttpResponseRedirect, QueryDict
+from django.shortcuts import resolve_url
+
+
+def redirect_to_login(next, login_url=None, redirect_field_name=REDIRECT_FIELD_NAME):
+    """
+    Redirect the user to the login page, passing the given 'next' page.
+    """
+    resolved_url = resolve_url(login_url or settings.LOGIN_URL)
+
+    login_url_parts = list(urlparse(resolved_url))
+    if redirect_field_name:
+        querystring = QueryDict(login_url_parts[4], mutable=True)
+        querystring[redirect_field_name] = next
+        login_url_parts[4] = querystring.urlencode(safe="/")
+
+    return HttpResponseRedirect(urlunparse(login_url_parts))
